Menambahkan DeathUI
