+++ Assets/Scripts/Data.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe73f316345247d439ecbf82a04d0c94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
